feat: improved discover and offers nav
Now the discover tab shows some dummy data about different places.

It's possible to click on "more button" or the places in the list view
to navigate to place detail. This page is empty and it will get data in
the future.

The offers nav has the same behavour.

Notice that is possible to navigate back by using a back-button on top
of the screen. This button works even if the browser is reoladed.

This commit adds the work nedded to achieve what is described above.