feat: add drawer navigation, auth and modal
This commit contains the following:

* Drawer Navigation

In some pages, a drawer is shown to the user.
This drawer contains 3 elements:

- Discover place to navigate to /places/tabs/discover
- Your bookings which is empty for now
- Auth to authenticate the user.

* Auth

For the auth part two new elements have been added:

-Auth service: handles the loging and logout of the user, currently
dummy implementation.
-Auth guard: To prevent the user to navigate through the app without
loging.

* Modal

This is used to present more information about a place when the user
clicks on the book button. Currently only shows the title of the place.
It also contains two buttons:

- Book. This will be used to book the place. Now close it but it
registers a "book" role
- Cancel. To close the modal. Registers a "close" role.