feat: add backend integration
The app now is able to connect to a backend in order to store
all the information shown to the user.

The chosen backed is Google Firebase and the comunication is performed
by using REST API

Please notice that the actual Firebase URL has been removed from
this commit for security purposes.

This commit contains all the changes to achieve the work described
above