feat: add Google Maps integration
Now the user is able to select the address of a place/offer by using
Google Maps.

By clicking on an specific place, its address and a picture of the
location will be presented to the user and stored in the backend.

The GoogleMaps API key and firebaseURL have been moved to the enviroment
folder.

This commit contains all the work described above.