chore: add camera integration
Now the user is able to pick an image by using the device camera or
an existing image file if the camera is not avaiable (or the user
doesn't want to use it).

In addition to this, a Firebase function has been added to upload the
image to the server. For this, some methods in new-offer and in
the place service were modified.

The API obtained from Firebase must be set in functions/index.js.
For security purposes, it hasn't been included in this commit.

This commit adds the work described above.